requirements.txt
pydantic-settings
pandas
google-generativeai
openpyxl